add `remove-eki` command
